Move VAOS unit tests: Batch 3
